Avoid setting `Error::Http` behind feature flags.
